Re-structruing the Project:- - seperate Serial Communication class in new package - seperate Speed Node in its package like Audio and Map - seperate Thread of Reading GPS Lines in same Package of Cpmmunication
